+++ tests/server/script.js
@@ -1,0 +1,6 @@
+function scriptJsTesting() {
+    document.querySelector("body").innerHTML += "Paragraph changed.";
+}
+document.addEventListener('DOMContentLoaded', function () {
+    scriptJsTesting()
+}, false);