+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog for _cewler_.
 
 ## Version 1.0.8
 + Added support for storing parsed URLs. Example command: cewler --output-urls urls.txt https://example.com
++ Added support for crawling JavaScript via both external files and tags. Example command: cewler --include-js https://example.com
 ## Version 1.0.7
 + Added crawling of error pages
 + Added crawling of HTML comments